add update account information client and server actions
