+++ src/routes/admin/accounts/components/edit-account/components/edit-email.svelte
@@ -1,24 +1,23 @@
 <script lang="ts">
-  import * as Dialog from '$lib/components/ui/dialog/index.js';
   import Loader from 'lucide-svelte/icons/loader';
   import * as Form from '$lib/components/ui/form/index.js';
   import { Input } from '$lib/components/ui/input/index.js';
-  import { editPasswordSchema, type EditPasswordSchema } from './schema';
+  import { editEmailSchema, type EditEmailSchema } from './schema';
   import { type SuperValidated, type Infer, superForm } from 'sveltekit-superforms';
   import { zodClient } from 'sveltekit-superforms/adapters';
   import { toast } from 'svelte-sonner';
-  import { useTableState } from '../table/state.svelte';
+  import { useTableState } from '../../table/state.svelte';
 
   interface Props {
-    editPasswordForm: SuperValidated<Infer<EditPasswordSchema>>;
+    editEmailForm: SuperValidated<Infer<EditEmailSchema>>;
   }
 
-  const { editPasswordForm }: Props = $props();
+  const { editEmailForm }: Props = $props();
 
   const tableState = useTableState();
 
-  const form = superForm(editPasswordForm, {
-    validators: zodClient(editPasswordSchema),
+  const form = superForm(editEmailForm, {
+    validators: zodClient(editEmailSchema),
     id: crypto.randomUUID(),
     onUpdate: ({ result }) => {
       const { status, data } = result;
@@ -40,41 +39,23 @@
   const { form: formData, enhance, submitting } = form;
 
   $effect(() => {
-    $formData.user_id = tableState.getActiveRow()?.user_id ?? '';
-
-    return () => {
-      form.reset();
-    };
+    if (tableState.showUpdate) {
+      $formData.user_id = tableState.getActiveRow()?.user_id ?? '';
+      $formData.email = tableState.getActiveRow()?.email ?? '';
+      return () => {
+        form.reset();
+      };
+    }
   });
 </script>
 
-<form method="POST" action="?/editSubjectEvent" use:enhance>
+<form method="POST" action="?/editEmailEvent" use:enhance>
   <input name="user_id" type="hidden" bind:value={$formData.user_id} />
-  <Form.Field {form} name="password">
+  <Form.Field {form} name="email">
     <Form.Control>
       {#snippet children({ props })}
-        <Form.Label>Password</Form.Label>
-        <Input
-          type="password"
-          {...props}
-          bind:value={$formData.password}
-          placeholder="Enter new password"
-        />
-      {/snippet}
-    </Form.Control>
-    <Form.FieldErrors />
-  </Form.Field>
-
-  <Form.Field {form} name="confirmPassword">
-    <Form.Control>
-      {#snippet children({ props })}
-        <Form.Label>Confirm Password</Form.Label>
-        <Input
-          type="password"
-          {...props}
-          bind:value={$formData.confirmPassword}
-          placeholder="Confirm new password"
-        />
+        <Form.Label>Email</Form.Label>
+        <Input {...props} bind:value={$formData.email} placeholder="Enter new email" />
       {/snippet}
     </Form.Control>
     <Form.FieldErrors />
@@ -87,7 +68,7 @@
           <Loader class="animate-spin" />
         </div>
       {/if}
-      Update Password
+      Update Email
     </Form.Button>
   </div>
 </form>
