+++ src/routes/admin/accounts/components/edit-account/components/edit-password.svelte
@@ -2,22 +2,22 @@
   import Loader from 'lucide-svelte/icons/loader';
   import * as Form from '$lib/components/ui/form/index.js';
   import { Input } from '$lib/components/ui/input/index.js';
-  import { editEmailSchema, type EditEmailSchema } from './schema';
+  import { editPasswordSchema, type EditPasswordSchema } from './schema';
   import { type SuperValidated, type Infer, superForm } from 'sveltekit-superforms';
   import { zodClient } from 'sveltekit-superforms/adapters';
   import { toast } from 'svelte-sonner';
-  import { useTableState } from '../table/state.svelte';
+  import { useTableState } from '../../table/state.svelte';
 
   interface Props {
-    editEmailForm: SuperValidated<Infer<EditEmailSchema>>;
+    editPasswordForm: SuperValidated<Infer<EditPasswordSchema>>;
   }
 
-  const { editEmailForm }: Props = $props();
+  const { editPasswordForm }: Props = $props();
 
   const tableState = useTableState();
 
-  const form = superForm(editEmailForm, {
-    validators: zodClient(editEmailSchema),
+  const form = superForm(editPasswordForm, {
+    validators: zodClient(editPasswordSchema),
     id: crypto.randomUUID(),
     onUpdate: ({ result }) => {
       const { status, data } = result;
@@ -39,23 +39,41 @@
   const { form: formData, enhance, submitting } = form;
 
   $effect(() => {
-    if (tableState.showUpdate) {
-      $formData.user_id = tableState.getActiveRow()?.user_id ?? '';
-      $formData.email = tableState.getActiveRow()?.email ?? '';
-      return () => {
-        form.reset();
-      };
-    }
+    $formData.user_id = tableState.getActiveRow()?.user_id ?? '';
+
+    return () => {
+      form.reset();
+    };
   });
 </script>
 
-<form method="POST" action="?/editSubjectEvent" use:enhance>
+<form method="POST" action="?/editPasswordEvent" use:enhance>
   <input name="user_id" type="hidden" bind:value={$formData.user_id} />
-  <Form.Field {form} name="email">
+  <Form.Field {form} name="password">
     <Form.Control>
       {#snippet children({ props })}
-        <Form.Label>Email</Form.Label>
-        <Input {...props} bind:value={$formData.email} placeholder="Enter new email" />
+        <Form.Label>Password</Form.Label>
+        <Input
+          type="password"
+          {...props}
+          bind:value={$formData.password}
+          placeholder="Enter new password"
+        />
+      {/snippet}
+    </Form.Control>
+    <Form.FieldErrors />
+  </Form.Field>
+
+  <Form.Field {form} name="confirmPassword">
+    <Form.Control>
+      {#snippet children({ props })}
+        <Form.Label>Confirm Password</Form.Label>
+        <Input
+          type="password"
+          {...props}
+          bind:value={$formData.confirmPassword}
+          placeholder="Confirm new password"
+        />
       {/snippet}
     </Form.Control>
     <Form.FieldErrors />
@@ -68,7 +86,7 @@
           <Loader class="animate-spin" />
         </div>
       {/if}
-      Update Email
+      Update Password
     </Form.Button>
   </div>
 </form>
